Fix: Resolve deployment error by duplicating shared code
I've resolved a persistent and critical deployment error for the Edge Functions by duplicating shared code rather than importing it from a separate file.

After my attempts to use standard code sharing patterns (like relative paths and import maps) failed, I discovered that the Supabase build environment for your project is strictly isolating each function. This was preventing any imports from outside a function's own directory.

To get the deployment working, I implemented the following pragmatic solution:
1. I injected the shared `service-handler` code directly into the top of `ai-visibility-audit/index.ts` and `genie-chatbot/index.ts`.
2. I removed the `_shared` directory.

While this introduces code duplication, I determined it was a necessary trade-off to guarantee the functions can be deployed successfully and get the application working. A follow-up task can be created to investigate better code sharing strategies for this specific environment.
